+++ src/tko/settings/rep_settings.py
@@ -65,14 +65,19 @@ class RepData:
     __tasks = "tasks"
     __flags = "flags"
     __lang = "lang"
+    __index = "index"
 
     defaults = {
         __expanded: [],
         __tasks: {},
         __flags: {},
         __new_items: [],
-        __lang: ""
+        __lang: "",
+        __index: 0
     }
+
+    def get_index(self) -> int:
+        return self.__get(RepData.__index)
 
     def get_expanded(self) -> list:
         return self.__get(RepData.__expanded)
@@ -103,6 +108,9 @@ class RepData:
     
     def set_lang(self, value: str):
         return self.__set(RepData.__lang, value)
+    
+    def set_index(self, value: int):
+        return self.__set(RepData.__index, value)
 
     def __get(self, key: str) -> Any:
         if key not in self.defaults: