+++ src/tko/util/runner.py
@@ -46,8 +46,11 @@ class Runner:
         
         term_print(Report.centralize(Sentence() + " " + cmd + " ", "─"))
         if cmd.startswith("node"):
-            term_print(Report.centralize(Sentence() + " Use Control-D caso precise finalizar a entrada ", "─"))
-
+            if os.name == "nt":
+                term_print(Report.centralize(Sentence() + " Use Control-Z Enter caso precise finalizar a entrada ", "─"))
+            else:
+                term_print(Report.centralize(Sentence() + " Use Control-D caso precise finalizar a entrada ", "─"))
+            
         answer = subprocess.run(cmd, shell=True, text=True)
         if answer.returncode != 0 and answer.returncode != 1:
             print(Runner.decode_code(answer.returncode))